Convert preferred model selection into hook
